requirements.txt
datasets
flask
flask_cors
flask_expects_json
requests
Pillow
gradio
Flask-APScheduler